backend/requirements.txt
flask
opencv-python
numpy
openai-whisper
pydub 